Updated text for not available and automate zip load

+++ lib/tasks/zip_load.rake
@@ -4,15 +4,19 @@ namespace :db do
   desc "Load Zip Codes into database from db directory." 
   task :zip_load => :environment do
     require 'csv'
-    ZipCode.delete_all
-    CSV.open("#{RAILS_ROOT}/db/zips.txt", "r") do |row|
-      zip = ZipCode.new
-      zip.zip = row[1]
-      zip.state = row[2]
-      zip.town = row[3]
-      zip.longitude = row[4]
-      zip.latitude = row[5]
-      zip.save!
+    require 'highline/import'       
+        
+    if agree("This task will destroy any Zip Codes in the database. Are you sure you want to \ncontinue? [yn] ")
+      ZipCode.delete_all
+      CSV.open("#{RAILS_ROOT}/db/zips.txt", "r") do |row|
+        zip = ZipCode.new
+        zip.zip = row[1]
+        zip.state = row[2]
+        zip.town = row[3]
+        zip.longitude = row[4]
+        zip.latitude = row[5]
+        zip.save!
+      end
     end
   end
 end
